% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{backsolveSNPeff}
\alias{backsolveSNPeff}
\title{Compute ridge-regression equivalent marker effects from GBLUP results ("Backsolve SNP effects")}
\usage{
backsolveSNPeff(Z, g)
}
\arguments{
\item{Z}{Centered marker matrix (dominance deviations must also be centered)}

\item{g}{The solutions (blups, i.e. GEBVs) from the GBLUP model}
}
\value{
matrix of SNP effects matching RR-BLUP / SNP-BLUP
}
\description{
From the GBLUP solutions and a centered SNP matrix backsolve SNP effects
}
\seealso{
Other helper: 
\code{\link{centerDosage}()},
\code{\link{crosses2predict}()},
\code{\link{dose2domDevGenotypic}()},
\code{\link{dose2domDev}()},
\code{\link{effectsArray2list}()},
\code{\link{genmap2recombfreq}()},
\code{\link{genoVarCovarMatFunc}()},
\code{\link{getAF}()},
\code{\link{getMAF}()},
\code{\link{getPropHom}()},
\code{\link{intensity}()},
\code{\link{kinship}()},
\code{\link{maf_filter}()},
\code{\link{quadform}()},
\code{\link{remove_invariant}()}
}
\concept{helper}
