% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{calcCrossLD}
\alias{calcCrossLD}
\title{Calculate the cross LD matrix based on both parents}
\usage{
calcCrossLD(sireID, damID, recombFreqMat, haploMat)
}
\arguments{
\item{sireID}{string, Male parent genotype ID.}

\item{damID}{string, Female parent genotype ID.}

\item{recombFreqMat}{a square symmetric matrix with values = (1-2*c1), where c1=matrix of expected recomb. frequencies. The choice to do 1-2c1 outside the function was made for computation efficiency; every operation on a big matrix takes time.}

\item{haploMat}{matrix of phased haplotypes, 2 rows per sample, cols = loci, {0,1}, rownames assumed to contain GIDs with a suffix, separated by "_" to distinguish haplotypes}
}
\value{
Potentially really large matrix representing the LD between loci in the cross
}
\description{
Wraps the \code{\link{calcGameticLD}} function. CrossLD = sireLD + damLD.
}
\details{
Columns of haploMat and row/cols of recombFreqMat should be in same order. May be worth computing in an R session using multi-threaded BLAS.
}
\seealso{
Other predCrossVar: 
\code{\link{calcGameticLD}()},
\code{\link{predCrossMeans}()},
\code{\link{predCrossVars}()}
}
\concept{predCrossVar}
