% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{centerDosage}
\alias{centerDosage}
\title{Centers dosage matrix}
\usage{
centerDosage(M)
}
\arguments{
\item{M}{dosage matrix. Assumes SNPs in M coded 0, 1, 2 (requires rounding dosages to integers). M is Nind x Mrow, numeric matrix, with row/columanes to indicate SNP/ind ID}
}
\value{

}
\description{
Centers dosage matrix, e.g. for use in whole-genome regressions like rrBLUP
}
\seealso{
Other helper: 
\code{\link{backsolveSNPeff}()},
\code{\link{crosses2predict}()},
\code{\link{dose2domDevGenotypic}()},
\code{\link{dose2domDev}()},
\code{\link{effectsArray2list}()},
\code{\link{genmap2recombfreq}()},
\code{\link{genoVarCovarMatFunc}()},
\code{\link{getAF}()},
\code{\link{getMAF}()},
\code{\link{getPropHom}()},
\code{\link{intensity}()},
\code{\link{kinship}()},
\code{\link{maf_filter}()},
\code{\link{quadform}()},
\code{\link{remove_invariant}()}
}
\concept{helper}
