% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDart2vcf.R
\name{convertDart2vcf}
\alias{convertDart2vcf}
\title{Convert DArTseqLD reports to VCF}
\usage{
convertDart2vcf(
  dartvcfInput,
  dartcountsInput,
  outName,
  nskipvcf = 2,
  nskipcounts = 3,
  ncores
)
}
\value{
The output of this step is written to disk as a \code{*.vcf.gz} completely compatible with downstream bioinformatics softwares!
}
\description{
This function converts a dual-file report format that we (NextGen Cassava Breeding) have arranged with DArT.
}
\details{
The output of this step is written to disk as a \code{*.vcf.gz} completely compatible with downstream bioinformatics softwares!

\strong{NOTICE:} This function is part of a family of functions (\code{"imputation_functions"}) developed as part of the NextGen Cassava Breeding Project genomic selection pipeline.
For some examples of their useage:
\itemize{
 \item \href{https://wolfemd.github.io/IITA_2021GS/}{IITA_2021GS (\strong{\emph{should be first use of this package}})}
 \item \href{https://wolfemd.github.io/NRCRI_2021GS/}{NRCRI_2021GS}
 \item \href{https://wolfemd.github.io/TARI_2020GS/}{TARI_2020GS}
}
}
\seealso{
Other imputation_functions: 
\code{\link{convertVCFtoDosage}()},
\code{\link{createGenomewideDosage}()},
\code{\link{filter_positions}()},
\code{\link{mergeVCFs}()},
\code{\link{postImputeFilterBeagle4pt1}()},
\code{\link{postImputeFilter}()},
\code{\link{runBeagle4pt1GL}()},
\code{\link{runBeagle5}()},
\code{\link{splitVCFbyChr}()}
}
\concept{imputation_functions}
