% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationPipeline.R
\name{convertVCFtoDosage}
\alias{convertVCFtoDosage}
\title{Convert a VCF file to dosage matrix}
\usage{
convertVCFtoDosage(pathIn, pathOut, vcfName)
}
\arguments{
\item{vcfName}{}
}
\description{
The function below will
}
\details{
(1) convert the input VCF to plink1.9 binary format and
(2) convert the plink binary to a dosage (0,1,2) matrix with special attention to which allele gets counted in the file.


Uses \href{https://www.cog-genomics.org/plink/}{\code{plink1.9}}. With \code{plink1.9}
there is some risk the counted allele could switch between
e.g. the reference panel and the progeny files because of allele freq. (see plink documentation).
To avoid this, went to extra trouble: write a file suffixed \code{*.alleleToCount}
listing SNP ID (column 1) and the ALT allele from the VCF (column 2).
Pass the file to \code{plink1.9} using the \code{--recode-allele} flag
to ensure all output dosages count the ALT allele consistent with the VCFs.
The reason to use \code{plink1.9} is that \code{Beagle5} imputed files
don't have a \strong{DS} (dosage) field that can be directly extracted.
Instead, phased genotypes e.g. \code{0|1} need to be converted to dosages
(e.g. \code{0|1 --> 1}, code{1|1 --> 2}).
An alternative might be to extract the haplotypes using code{vcftools} and
manually compute the dosages.

\strong{NOTICE:} This function is part of a family of functions (\code{"imputation_functions"}) developed as part of the NextGen Cassava Breeding Project genomic selection pipeline.
For some examples of their useage:
\itemize{
 \item \href{https://wolfemd.github.io/IITA_2021GS/}{IITA_2021GS (\strong{\emph{should be first use of this package}})}
 \item \href{https://wolfemd.github.io/NRCRI_2021GS/}{NRCRI_2021GS}
 \item \href{https://wolfemd.github.io/TARI_2020GS/}{TARI_2020GS}
}
}
\seealso{
Other imputation_functions: 
\code{\link{convertDart2vcf}()},
\code{\link{createGenomewideDosage}()},
\code{\link{filter_positions}()},
\code{\link{mergeVCFs}()},
\code{\link{postImputeFilterBeagle4pt1}()},
\code{\link{postImputeFilter}()},
\code{\link{runBeagle4pt1GL}()},
\code{\link{runBeagle5}()},
\code{\link{splitVCFbyChr}()}
}
\concept{imputation_functions}
