% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{crosses2predict}
\alias{crosses2predict}
\title{Make a data.frame of pairwise crosses between a set parents}
\usage{
crosses2predict(parents)
}
\arguments{
\item{parents}{}
}
\value{
tibble, two columns, sireID and damID, all pairwise crosses (see details).
}
\description{
Make a data.frame of all pairwise matings given a vector of parent IDs.
Include selfs. No reciprocal crosses, i.e. use as male == use as female.
Diagonal and upper-triangle of mating matrix.
}
\seealso{
Other helper: 
\code{\link{backsolveSNPeff}()},
\code{\link{centerDosage}()},
\code{\link{dose2domDevGenotypic}()},
\code{\link{dose2domDev}()},
\code{\link{effectsArray2list}()},
\code{\link{genmap2recombfreq}()},
\code{\link{genoVarCovarMatFunc}()},
\code{\link{getAF}()},
\code{\link{getMAF}()},
\code{\link{getPropHom}()},
\code{\link{intensity}()},
\code{\link{kinship}()},
\code{\link{maf_filter}()},
\code{\link{quadform}()},
\code{\link{remove_invariant}()}
}
\concept{helper}
