% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassavabasePipeline.R
\name{detectExptDesigns}
\alias{detectExptDesigns}
\title{"Detect" experimental designs}
\usage{
detectExptDesigns(indata)
}
\arguments{
\item{indata}{}
}
\value{
Returns the input data.frame with two new columns indicating TRUE/FALSE, for each trial (location-year-studyName),
whether the trial \code{replicate / repInTrial} columns indicates \strong{\code{CompleteBlocks}} and
if the \code{blockInRep / blockNumber} contains \strong{\code{IncompleteBlocks}}.
}
\description{
After cleaning up cassavabase trial data, the next step is to run some \emph{ad hoc} code that checks the experimental design of each trial.
If you are absolutely certain of the usage of the design variables in your dataset, you might not need this step.
}
\details{
Returns the input data.frame with two new columns indicating TRUE/FALSE, for each trial (location-year-studyName),
whether the trial \code{replicate / repInTrial} columns indicates \strong{\code{CompleteBlocks}} and
if the \code{blockInRep / blockNumber} contains \strong{\code{IncompleteBlocks}}.

Examples of reasons to do the step below:
\itemize{
 \item Some trials appear to be complete blocked designs and the blockNumber is used instead of replicate, which is what most use.
 \item Some complete block designs have nested, incomplete sub-blocks, others simply copy the "replicate" variable into the "blockNumber variable"
 \item  Some trials have only incomplete blocks \emph{but} the incomplete block info might be in the replicate \emph{and/or} the blockNumber column
}

One reason it might be important to get this right is that the variance among complete blocks might not be the same among incomplete blocks. If we treat a mixture of complete and incomplete blocks as part of the same random-effect (replicated-within-trial), we assume they have the same variance.

\strong{NOTICE:} This function is part of a family of functions (\code{"cassavabase_pheno_pipeline"}) developed as part of the NextGen Cassava Breeding Project genomic selection pipeline.
For some examples of their useage:
\itemize{
 \item \href{https://wolfemd.github.io/IITA_2021GS/}{IITA_2021GS (\strong{\emph{should be first use of this package}})}
 \item \href{https://wolfemd.github.io/NRCRI_2021GS/}{NRCRI_2021GS}
 \item \href{https://wolfemd.github.io/TARI_2020GS/}{TARI_2020GS}
 \item \href{https://wolfemd.github.io/IITA_2020GS/}{IITA_2020GS}
}
}
\seealso{
Other cassavabase_pheno_pipeline: 
\code{\link{calcPropMissing}()},
\code{\link{curateTrialOneTrait}()},
\code{\link{curateTrialsByTrait}()},
\code{\link{fitMultiTrialModel}()},
\code{\link{makeTrialTypeVar}()},
\code{\link{nestByTrials}()},
\code{\link{nestDesignsDetectedByTraits}()},
\code{\link{nestForMultiTrialAnalysis}()},
\code{\link{nestTrialsByTrait}()},
\code{\link{readDBdata}()},
\code{\link{renameAndSelectCols}()}
}
\concept{cassavabase_pheno_pipeline}
