% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{effectsArray2list}
\alias{effectsArray2list}
\title{Converts an array of posterior samples of multi-trait marker effects to a named list (one for each trait).}
\usage{
effectsArray2list(effectsArray, snpIDs, traits, nIter, burnIn, thin)
}
\arguments{
\item{effectsArray}{According to BGLR documentation: 3D array, with dim=c(nRow,p,traits), where nRow number of MCMC samples saved, p is the number of predictors and traits is the number of traits. \code{\link[BGLR]{Multitrait}}writes a binary file to disk when saveEffects=TRUE is specified. It can be read to R with \code{\link[BGLR]{readBinMatMultitrait}}.}

\item{snpIDs}{character vector with labels for the predictors (SNPs), numeric should work too, but untested.}

\item{traits}{character vector to label the traits.}

\item{nIter}{number of iterations used for MCMC; used internally only to exclude burn-in samples from computation}

\item{burnIn}{burnIn for MCMC; used internally only to exclude burn-in samples from computation}

\item{thin}{thin for MCMC; used internally only to exclude burn-in samples from computation}
}
\value{
list of matrices, one matrix per trait, each matrix has \code{nrow((nIter-burnIn)/thin)} and \code{ncol(length(snpIDs))}. Each element of the list is named with a string identifying the trait and the colnames of each matrix are labelled with snpIDs.
}
\description{
Converts an array of posterior samples of multi-trait marker effects to a named list (one for each trait).
}
\seealso{
Other helper: 
\code{\link{backsolveSNPeff}()},
\code{\link{centerDosage}()},
\code{\link{crosses2predict}()},
\code{\link{dose2domDevGenotypic}()},
\code{\link{dose2domDev}()},
\code{\link{genmap2recombfreq}()},
\code{\link{genoVarCovarMatFunc}()},
\code{\link{getAF}()},
\code{\link{getMAF}()},
\code{\link{getPropHom}()},
\code{\link{intensity}()},
\code{\link{kinship}()},
\code{\link{maf_filter}()},
\code{\link{quadform}()},
\code{\link{remove_invariant}()}

Other helper: 
\code{\link{backsolveSNPeff}()},
\code{\link{centerDosage}()},
\code{\link{crosses2predict}()},
\code{\link{dose2domDevGenotypic}()},
\code{\link{dose2domDev}()},
\code{\link{genmap2recombfreq}()},
\code{\link{genoVarCovarMatFunc}()},
\code{\link{getAF}()},
\code{\link{getMAF}()},
\code{\link{getPropHom}()},
\code{\link{intensity}()},
\code{\link{kinship}()},
\code{\link{maf_filter}()},
\code{\link{quadform}()},
\code{\link{remove_invariant}()}
}
\concept{helper}
