% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationPipeline.R
\name{filter_positions}
\alias{filter_positions}
\title{Subset a VCF file based on provided positions}
\usage{
filter_positions(inPath = NULL, inVCF, positionFile, outPath = NULL, outName)
}
\arguments{
\item{positionFile}{see VCFtools}

\item{outName}{}
}
\value{

}
\description{
Uses \href{https://vcftools.github.io/index.html}{\code{vcftools --positions}}
}
\details{
\strong{NOTICE:} This function is part of a family of functions (\code{"imputation_functions"}) developed as part of the NextGen Cassava Breeding Project genomic selection pipeline.
For some examples of their useage:
\itemize{
 \item \href{https://wolfemd.github.io/IITA_2021GS/}{IITA_2021GS (\strong{\emph{should be first use of this package}})}
 \item \href{https://wolfemd.github.io/NRCRI_2021GS/}{NRCRI_2021GS}
 \item \href{https://wolfemd.github.io/TARI_2020GS/}{TARI_2020GS}
}
}
\seealso{
Other imputation_functions: 
\code{\link{convertDart2vcf}()},
\code{\link{convertVCFtoDosage}()},
\code{\link{createGenomewideDosage}()},
\code{\link{mergeVCFs}()},
\code{\link{postImputeFilterBeagle4pt1}()},
\code{\link{postImputeFilter}()},
\code{\link{runBeagle4pt1GL}()},
\code{\link{runBeagle5}()},
\code{\link{splitVCFbyChr}()}
}
\concept{imputation_functions}
