% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{genmap2recombfreq}
\alias{genmap2recombfreq}
\title{Create a matrix of pairwise recombination frequencies from a genetic map}
\usage{
genmap2recombfreq(m, nChr)
}
\arguments{
\item{m}{vector of centiMorgan-scale genetic positions. names(m) correspond to a SNP_ID. Since m potentially contains all chromosomes, sets recomb. freq. b/t chrom. to 0.5}

\item{nChr}{number of chromosomes}
}
\value{
potentially really large matrix of pairwise recombination frequencies between loci
}
\description{
Compute the pairwise recombination frequencies between all loci from genetic map positions.
}
\details{
names(m) must be formatted as "chr"_"id" with "chr" being integer. For example: 2_QTL1 for a locus on chr. 2.
May be worth computing in an R session using multi-threaded BLAS.
}
\seealso{
Other helper: 
\code{\link{backsolveSNPeff}()},
\code{\link{centerDosage}()},
\code{\link{crosses2predict}()},
\code{\link{dose2domDevGenotypic}()},
\code{\link{dose2domDev}()},
\code{\link{effectsArray2list}()},
\code{\link{genoVarCovarMatFunc}()},
\code{\link{getAF}()},
\code{\link{getMAF}()},
\code{\link{getPropHom}()},
\code{\link{intensity}()},
\code{\link{kinship}()},
\code{\link{maf_filter}()},
\code{\link{quadform}()},
\code{\link{remove_invariant}()}
}
\concept{helper}
