% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicMateSelectR-package.R
\docType{package}
\name{genomicMateSelectR-package}
\alias{genomicMateSelectR}
\alias{genomicMateSelectR-package}
\title{genomicMateSelectR: Genomic Mate Selection}
\description{
Predict the usefulness of a cross. Multi-trait selection index mean and variance prediction based on parental haplotypes, a genetic map, and marker effects. Implement cross-validation procedures to estimate the accuracy predicting the performance (breeding value, total genetic value including dominance) of previously unobserved individuals as well as cross means and variances. Successor package to predCrossVar.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wolfemd/genomicMateSelectR}
  \item Report bugs at \url{https://github.com/wolfemd/genomicMateSelectR/issues}
}

}
\keyword{internal}
