% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{genoVarCovarMatFunc}
\alias{genoVarCovarMatFunc}
\title{Compute the LD matrix from dosages}
\usage{
genoVarCovarMatFunc(Z)
}
\arguments{
\item{Z}{column-centered matrix of SNP dosages. Assumes SNPs in Z were originally coded 0, 1, 2 were column centered.}
}
\value{
NOTE: this matrix is going to be big in practice.
 The \eqn{p_{SNP} \times p_{SNP}} variance-covariance matrix of SNP dosages.
 may be worth computing in an R session using multi-threaded BLAS
}
\description{
Compute the \eqn{p_{SNP} \times p_{SNP}} variance-covariance matrix of SNP dosages.
This is an estimator of the LD between loci within a given population.
}
\seealso{
Other helper: 
\code{\link{backsolveSNPeff}()},
\code{\link{centerDosage}()},
\code{\link{crosses2predict}()},
\code{\link{dose2domDevGenotypic}()},
\code{\link{dose2domDev}()},
\code{\link{effectsArray2list}()},
\code{\link{genmap2recombfreq}()},
\code{\link{getAF}()},
\code{\link{getMAF}()},
\code{\link{getPropHom}()},
\code{\link{intensity}()},
\code{\link{kinship}()},
\code{\link{maf_filter}()},
\code{\link{quadform}()},
\code{\link{remove_invariant}()}
}
\concept{helper}
