% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentWiseCrossVal.R
\name{getMarkEffs}
\alias{getMarkEffs}
\title{Title}
\usage{
getMarkEffs(
  parentfolds,
  blups,
  gid,
  modelType,
  grms,
  dosages,
  ncores,
  nBLASthreads = NULL
)
}
\arguments{
\item{blups}{nested data.frame with list-column "TrainingData" containing}

\item{nBLASthreads}{}
}
\value{

}
\description{
Title
}
