% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getPropHom}
\alias{getPropHom}
\title{Compute the per-individual proportion homozygous}
\usage{
getPropHom(M)
}
\arguments{
\item{M}{dosage matrix. Assumes SNPs in M coded 0, 1, 2 (requires rounding dosages to integers). M is Nind x Mrow, numeric matrix, with row/columanes to indicate SNP/ind ID}
}
\value{
vector of per-individual proportion homozygous
}
\description{
Compute the per-individual proportion homozygous.
For example, to use as a predictor for a directional dominance model.
}
\seealso{
Other helper: 
\code{\link{backsolveSNPeff}()},
\code{\link{centerDosage}()},
\code{\link{crosses2predict}()},
\code{\link{dose2domDevGenotypic}()},
\code{\link{dose2domDev}()},
\code{\link{effectsArray2list}()},
\code{\link{genmap2recombfreq}()},
\code{\link{genoVarCovarMatFunc}()},
\code{\link{getAF}()},
\code{\link{getMAF}()},
\code{\link{intensity}()},
\code{\link{kinship}()},
\code{\link{maf_filter}()},
\code{\link{quadform}()},
\code{\link{remove_invariant}()}
}
\concept{helper}
