% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{intensity}
\alias{intensity}
\title{Standardized selection intensity}
\usage{
intensity(propSel)
}
\arguments{
\item{propSel}{proportion selection}
}
\value{

}
\description{
Compute the standardized selection intensity from the proportion selection.
}
\seealso{
Other helper: 
\code{\link{backsolveSNPeff}()},
\code{\link{centerDosage}()},
\code{\link{crosses2predict}()},
\code{\link{dose2domDevGenotypic}()},
\code{\link{dose2domDev}()},
\code{\link{effectsArray2list}()},
\code{\link{genmap2recombfreq}()},
\code{\link{genoVarCovarMatFunc}()},
\code{\link{getAF}()},
\code{\link{getMAF}()},
\code{\link{getPropHom}()},
\code{\link{kinship}()},
\code{\link{maf_filter}()},
\code{\link{quadform}()},
\code{\link{remove_invariant}()}
}
\concept{helper}
