% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentWiseCrossVal.R
\name{makeParentFolds}
\alias{makeParentFolds}
\title{Make parent-wise cross-validation folds}
\usage{
makeParentFolds(ped, gid, nrepeats = 5, nfolds = 5, seed = NULL)
}
\arguments{
\item{ped}{data.frame, 3 columns, "GID" (or \code{gid}), "sireID", "damID"
for male and female parent, respectively.}

\item{gid}{string variable name used for genotype ID's in e.g. \code{blups} (default="GID")}

\item{nrepeats}{number of repeats}

\item{nfolds}{number of folds,}

\item{seed}{integer, make reproducible}
}
\description{
Create parent-wise cross-validation folds based on an input pedigree.
}
