% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentWiseCrossVal.R
\name{meanPredAccuracy}
\alias{meanPredAccuracy}
\title{Title}
\usage{
meanPredAccuracy(
  crossValOut,
  snpeffs,
  ped,
  modelType,
  selInd = FALSE,
  SIwts = NULL
)
}
\arguments{
\item{SIwts}{}
}
\description{
Title
}
