% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationPipeline.R
\name{mergeVCFs}
\alias{mergeVCFs}
\title{Merge two VCFs into one}
\usage{
mergeVCFs(inPath = NULL, inVCF1, inVCF2, outPath = NULL, outName)
}
\arguments{
\item{outName}{}
}
\value{

}
\description{
Uses \href{https://samtools.github.io/bcftools/bcftools.html}{\code{bcftools merge}}. Also requires \code{tabix}.
}
\details{
\strong{NOTICE:} This function is part of a family of functions (\code{"imputation_functions"}) developed as part of the NextGen Cassava Breeding Project genomic selection pipeline.
For some examples of their useage:
\itemize{
 \item \href{https://wolfemd.github.io/IITA_2021GS/}{IITA_2021GS (\strong{\emph{should be first use of this package}})}
 \item \href{https://wolfemd.github.io/NRCRI_2021GS/}{NRCRI_2021GS}
 \item \href{https://wolfemd.github.io/TARI_2020GS/}{TARI_2020GS}
}
}
\seealso{
Other imputation_functions: 
\code{\link{convertDart2vcf}()},
\code{\link{convertVCFtoDosage}()},
\code{\link{createGenomewideDosage}()},
\code{\link{filter_positions}()},
\code{\link{postImputeFilterBeagle4pt1}()},
\code{\link{postImputeFilter}()},
\code{\link{runBeagle4pt1GL}()},
\code{\link{runBeagle5}()},
\code{\link{splitVCFbyChr}()}
}
\concept{imputation_functions}
