% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassavabasePipeline.R
\name{nestByTrials}
\alias{nestByTrials}
\title{Nest data.frame by trials}
\usage{
nestByTrials(indata)
}
\arguments{
\item{indata}{}
}
\value{

}
\description{
Also, Create some explicitly nested variables including loc and year to nest with the trial data
}
\details{
\strong{NOTICE:} This function is part of a family of functions (\code{"cassavabase_pheno_pipeline"}) developed as part of the NextGen Cassava Breeding Project genomic selection pipeline.
For some examples of their useage:
\itemize{
 \item \href{https://wolfemd.github.io/IITA_2021GS/}{IITA_2021GS (\strong{\emph{should be first use of this package}})}
 \item \href{https://wolfemd.github.io/NRCRI_2021GS/}{NRCRI_2021GS}
 \item \href{https://wolfemd.github.io/TARI_2020GS/}{TARI_2020GS}
 \item \href{https://wolfemd.github.io/IITA_2020GS/}{IITA_2020GS}
}
}
\seealso{
Other cassavabase_pheno_pipeline: 
\code{\link{calcPropMissing}()},
\code{\link{curateTrialOneTrait}()},
\code{\link{curateTrialsByTrait}()},
\code{\link{detectExptDesigns}()},
\code{\link{fitMultiTrialModel}()},
\code{\link{makeTrialTypeVar}()},
\code{\link{nestDesignsDetectedByTraits}()},
\code{\link{nestForMultiTrialAnalysis}()},
\code{\link{nestTrialsByTrait}()},
\code{\link{readDBdata}()},
\code{\link{renameAndSelectCols}()}
}
\concept{cassavabase_pheno_pipeline}
