% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassavabasePipeline.R
\name{nestDesignsDetectedByTraits}
\alias{nestDesignsDetectedByTraits}
\title{Set-up for a by-trait analysis downstream}
\usage{
nestDesignsDetectedByTraits(indata, traits)
}
\arguments{
\item{indata}{data.frame with experimental designs detected, output from \code{\link{detectExptDesigns}}}

\item{traits}{character vec, names of traits to include}
}
\description{
Need to restructure the data from per-trial by regrouping by trait.
}
\details{
\strong{NOTICE:} This function is part of a family of functions (\code{"cassavabase_pheno_pipeline"}) developed as part of the NextGen Cassava Breeding Project genomic selection pipeline.
For some examples of their useage:
\itemize{
 \item \href{https://wolfemd.github.io/IITA_2021GS/}{IITA_2021GS (\strong{\emph{should be first use of this package}})}
 \item \href{https://wolfemd.github.io/NRCRI_2021GS/}{NRCRI_2021GS}
 \item \href{https://wolfemd.github.io/TARI_2020GS/}{TARI_2020GS}
 \item \href{https://wolfemd.github.io/IITA_2020GS/}{IITA_2020GS}
}
}
\seealso{
Other cassavabase_pheno_pipeline: 
\code{\link{calcPropMissing}()},
\code{\link{curateTrialOneTrait}()},
\code{\link{curateTrialsByTrait}()},
\code{\link{detectExptDesigns}()},
\code{\link{fitMultiTrialModel}()},
\code{\link{makeTrialTypeVar}()},
\code{\link{nestByTrials}()},
\code{\link{nestForMultiTrialAnalysis}()},
\code{\link{nestTrialsByTrait}()},
\code{\link{readDBdata}()},
\code{\link{renameAndSelectCols}()}
}
\concept{cassavabase_pheno_pipeline}
