% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationPipeline.R
\name{postImputeFilter}
\alias{postImputeFilter}
\title{Apply quality filters after imputation by Beagle5.0}
\usage{
postImputeFilter(
  inPath = NULL,
  inName,
  outPath = NULL,
  outName,
  DR2thresh = 0.75,
  HWEthresh = 1e-20,
  MAFthresh = 0.005
)
}
\arguments{
\item{inPath}{path to input VCFs}

\item{inName}{imputed input VCF name}

\item{outPath}{path for output to be written to}

\item{outName}{name desired, don't add file suffix, will be *.vcf.gz}

\item{DR2thresh}{remove if imputation quality score \code{DR2} supplied in the VCF INFO field by Beagle is less than or equal to the threshold}

\item{HWEthresh}{remove if p-value HWE from \code{vcftools --hardy} less than threshold, smaller p-value means more departure from HWE}

\item{MAFthresh}{remove if minor allele frequency less than threshold}
}
\description{
Remove markers from the imputed VCF based on adjustable thresholds.
For VCFs imputed by Beagle5
}
\details{
Uses \href{https://vcftools.github.io/index.html}{vcftools} and R.
of minor allee frequency \code{MAFthresh}, p-value measuring likelihood of divergence from hardy-weinberg equilibrium ()

\strong{NOTICE:} This function is part of a family of functions (\code{"imputation_functions"}) developed as part of the NextGen Cassava Breeding Project genomic selection pipeline.
For some examples of their useage:
\itemize{
 \item \href{https://wolfemd.github.io/IITA_2021GS/}{IITA_2021GS (\strong{\emph{should be first use of this package}})}
 \item \href{https://wolfemd.github.io/NRCRI_2021GS/}{NRCRI_2021GS}
 \item \href{https://wolfemd.github.io/TARI_2020GS/}{TARI_2020GS}
}
}
\seealso{
Other imputation_functions: 
\code{\link{convertDart2vcf}()},
\code{\link{convertVCFtoDosage}()},
\code{\link{createGenomewideDosage}()},
\code{\link{filter_positions}()},
\code{\link{mergeVCFs}()},
\code{\link{postImputeFilterBeagle4pt1}()},
\code{\link{runBeagle4pt1GL}()},
\code{\link{runBeagle5}()},
\code{\link{splitVCFbyChr}()}
}
\concept{imputation_functions}
