% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmsFunctions.R
\name{predictCrosses}
\alias{predictCrosses}
\title{Predict crosses}
\usage{
predictCrosses(
  modelType,
  stdSelInt = 2.67,
  selInd,
  SIwts = NULL,
  CrossesToPredict,
  snpeffs,
  dosages,
  haploMat,
  recombFreqMat,
  ncores = 1,
  nBLASthreads = NULL,
  predTheMeans = TRUE,
  predTheVars = TRUE
)
}
\arguments{
\item{modelType}{string, A, AD or DirDom. A and AD representing model with}

\item{selInd}{logical, TRUE/FALSE, selection index accuracy estimates,
requires input weights via \code{SIwts}}

\item{CrossesToPredict}{data.frame or tibble, col/colnames: sireID, damID. sireID and damID must both be in the haploMat.}

\item{snpeffs}{the element \code{genomicPredOut[[1]]} of the output of
\code{\link{runGenomicPredictions}}.}

\item{dosages}{dosage matrix. required only for modelType=="DirDom".
Assumes SNPs coded 0, 1, 2. Nind rows x Nsnp
cols, numeric matrix, with rownames and colnames to indicate SNP/ind ID}

\item{haploMat}{matrix of phased haplotypes, 2 rows per sample, cols = loci, {0,1},
rownames assumed to contain GIDs with a suffix, separated by "_" to distinguish haplotypes.
Currently, the haplotypes must be distinguished by the mandatory suffixes "_HapA" and "_HapB".}

\item{recombFreqMat}{a square symmetric matrix with values = (1-2*c1), where c1=matrix of expected recomb. frequencies. The choice to do 1-2c1 outside the function was made for computation efficiency; every operation on a big matrix takes time.}

\item{ncores}{number of cores}

\item{nBLASthreads}{number of cores for each worker to use for multi-thread BLAS}

\item{predTheMeans}{default: TRUE, t/f whether to predict cross means}

\item{predTheVars}{default: TRUE, t/f whether to predict cross vars}
}
\value{
tibble, one row, two list columns (basically a named two-element
list of lists): \code{tidyPreds[[1]]} and \code{rawPreds[[1]]}.
code{tidyPreds[[1]]}: tidy output, fewer details. sireID, damID, Nsegsnps, predOf,Trait, predMean, predVar, predSD, predUsefulnesstibble of predicted GEBV/GETGV, all traits and potentially
SELIND genomic BLUPs along the columns. \code{rawPreds[[1]]}: more detailed output, list of 2 ("predMeans" tibble and "predVars" tibble).
}
\description{
Predict potentially for multiple traits, the means, variances and
trait-trait covariances in a set ofuser supplied crosses.l
If requested, computed the selection index means and variances.
Computes the usefulness criteria \eqn{UC_{parent}} and \eqn{UC_{variety}}
potentially with a user supplied standardized selection intensity value
\code{stdSelInt}. Output enables easy ranking of potential crosses.
This function takes the matrices of snpeffects output
(\code{genomicPredOut[[1]]}) from the \code{\link{runGenomicPredictions}}
function (when \code{getMarkEffs=TRUE}).
This is a wrapper function around \code{\link{predCrossVars} and
\link{predCrossMeans}}.
}
\seealso{
Other prediction_functions: 
\code{\link{runGenomicPredictions}()}
}
\concept{prediction_functions}
