% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentWiseCrossVal.R
\name{predictCrossVars}
\alias{predictCrossVars}
\title{Title}
\usage{
predictCrossVars(
  modelType,
  snpeffs,
  parentfolds,
  haploMat,
  recombFreqMat,
  ncores,
  nBLASthreads = NULL
)
}
\arguments{
\item{nBLASthreads}{}
}
\description{
Title
}
