% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{predOneCross}
\alias{predOneCross}
\title{Title}
\usage{
predOneCross(
  sireID,
  damID,
  modelType,
  haploMat,
  recombFreqMat,
  predType,
  postMeanAddEffects,
  postMeanDomEffects = NULL,
  AddEffectList = NULL,
  DomEffectList = NULL,
  nBLASthreads,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
