% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{predOneCrossVar}
\alias{predOneCrossVar}
\title{Title}
\usage{
predOneCrossVar(
  Trait1,
  Trait2,
  progenyLD,
  modelType,
  predType,
  postMeanAddEffects,
  postMeanDomEffects = NULL,
  AddEffectList = NULL,
  DomEffectList = NULL,
  segsnps2keep = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
