% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{quadform}
\alias{quadform}
\title{Quadratic matrix form}
\usage{
quadform(D, x, y)
}
\arguments{
\item{D}{square symmetric matrix}

\item{x}{vector 1}

\item{y}{vector 2}
}
\value{
square symmetrix matrix
}
\description{
does \eqn{x \%\ast\% D \%\ast\% t(x)} a bit faster
}
\seealso{
Other helper: 
\code{\link{backsolveSNPeff}()},
\code{\link{centerDosage}()},
\code{\link{crosses2predict}()},
\code{\link{dose2domDevGenotypic}()},
\code{\link{dose2domDev}()},
\code{\link{effectsArray2list}()},
\code{\link{genmap2recombfreq}()},
\code{\link{genoVarCovarMatFunc}()},
\code{\link{getAF}()},
\code{\link{getMAF}()},
\code{\link{getPropHom}()},
\code{\link{intensity}()},
\code{\link{kinship}()},
\code{\link{maf_filter}()},
\code{\link{remove_invariant}()}
}
\concept{helper}
