% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationPipeline.R
\name{runBeagle4pt1GL}
\alias{runBeagle4pt1GL}
\title{Run Beagle4.1 to impute a target VCF with a reference VCF}
\usage{
runBeagle4pt1GL(
  targetVCF,
  refVCF,
  mapFile,
  outName,
  nthreads,
  maxmem = "500g",
  impute = TRUE,
  ne = 1e+05,
  samplesToExclude = NULL,
  niter = 10
)
}
\arguments{
\item{targetVCF}{passes to Beagle `\code{gl=targetVCF}`}

\item{refVCF}{passes to Beagle `\code{ref=targetVCF}`}

\item{mapFile}{passes to Beagle `\code{map=mapFile}`}

\item{outName}{passes to Beagle `\code{out=outName} (\strong{don't put file
suffix, Beagle adds \code{*.vcf.gz}}).}

\item{nthreads}{passes to Beagle `\code{nthreads=nthreads}`}

\item{maxmem}{passes to java `\code{-Xmx<maxmem>}`}

\item{impute}{passes to Beagle `\code{impute=TRUE}`}

\item{ne}{passes to Beagle `\code{ne=ne}`}

\item{niter}{passes to Beagle `\code{niterations=niter}`}
}
\value{

}
\description{
\href{https://faculty.washington.edu/browning/beagle/b4_1.html}{Beagle4.1} is
slower than Beagle5 by far. However, it can use genotype-likelihoods (the GL
VCF field) for its algorithm, which is potentially more accurate. In cases
like imputing GBS data, or if you have time to wait around, this should be
better than Beagle5.0 for a first pass imputation of observed sites. Impute
with \href{https://faculty.washington.edu/browning/beagle/b5_0.html}{Beagle
V5.0}. Use an "imputation reference panel". Refer to Beagle documentation for
meaning of arguments passed.
}
\details{
\strong{NOTICE:} This function is part of a family of functions (\code{"imputation_functions"}) developed as part of the NextGen Cassava Breeding Project genomic selection pipeline.
For some examples of their useage:
\itemize{
 \item \href{https://wolfemd.github.io/IITA_2021GS/}{IITA_2021GS (\strong{\emph{should be first use of this package}})}
 \item \href{https://wolfemd.github.io/NRCRI_2021GS/}{NRCRI_2021GS}
 \item \href{https://wolfemd.github.io/TARI_2020GS/}{TARI_2020GS}
}
}
\seealso{
Other imputation_functions: 
\code{\link{convertDart2vcf}()},
\code{\link{convertVCFtoDosage}()},
\code{\link{createGenomewideDosage}()},
\code{\link{filter_positions}()},
\code{\link{mergeVCFs}()},
\code{\link{postImputeFilterBeagle4pt1}()},
\code{\link{postImputeFilter}()},
\code{\link{runBeagle5}()},
\code{\link{splitVCFbyChr}()}
}
\concept{imputation_functions}
