% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationPipeline.R
\name{runBeagle5}
\alias{runBeagle5}
\title{Run Beagle5 to impute a target VCF with a reference VCF}
\usage{
runBeagle5(
  targetVCF,
  refVCF = NULL,
  mapFile,
  outName,
  nthreads,
  maxmem = "500g",
  impute = TRUE,
  ne = 1e+05,
  samplesToExclude = NULL
)
}
\arguments{
\item{targetVCF}{passes to Beagle `\code{gt=targetVCF}`}

\item{refVCF}{passes to Beagle `\code{ref=refVCF}`}

\item{mapFile}{passes to Beagle `\code{map=mapFile}`}

\item{outName}{passes to Beagle `\code{out=outName} (\strong{don't put file suffix, Beagle adds \code{*.vcf.gz}}).}

\item{nthreads}{passes to Beagle `\code{nthreads=nthreads}`}

\item{maxmem}{passes to java `\code{-Xmx<maxmem>}`}

\item{impute}{passes to Beagle `\code{impute=TRUE}`}

\item{ne}{passes to Beagle `\code{ne=ne}`}

\item{samplesToExclude}{}
}
\value{

}
\description{
Impute with \href{https://faculty.washington.edu/browning/beagle/b5_0.html}{Beagle V5.0}. Use an "imputation reference panel".
Refer to Beagle documentation for meaning of arguments passed.
}
\details{
\strong{NOTICE:} This function is part of a family of functions (\code{"imputation_functions"}) developed as part of the NextGen Cassava Breeding Project genomic selection pipeline.

\code{java -Xms2g -Xmx [maxmem] -jar /programs/beagle/beagle.jar gt= [targetVCF] map= [mapFile] ref= [refVCF] out= [outName] nthreads= [nthreads] impute= [impute]  ne=  [ne]}

For some examples of their useage:
\itemize{
 \item \href{https://wolfemd.github.io/IITA_2021GS/}{IITA_2021GS (\strong{\emph{should be first use of this package}})}
 \item \href{https://wolfemd.github.io/NRCRI_2021GS/}{NRCRI_2021GS}
 \item \href{https://wolfemd.github.io/TARI_2020GS/}{TARI_2020GS}
}
}
\seealso{
Other imputation_functions: 
\code{\link{convertDart2vcf}()},
\code{\link{convertVCFtoDosage}()},
\code{\link{createGenomewideDosage}()},
\code{\link{filter_positions}()},
\code{\link{mergeVCFs}()},
\code{\link{postImputeFilterBeagle4pt1}()},
\code{\link{postImputeFilter}()},
\code{\link{runBeagle4pt1GL}()},
\code{\link{splitVCFbyChr}()}
}
\concept{imputation_functions}
