% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmsFunctions.R
\name{runGenomicPredictions}
\alias{runGenomicPredictions}
\title{Run genomic predictions}
\usage{
runGenomicPredictions(
  modelType,
  selInd,
  SIwts = NULL,
  getMarkEffs = FALSE,
  returnPEV = FALSE,
  blups,
  grms,
  dosages = NULL,
  gid = "GID",
  ncores = 1,
  nBLASthreads = NULL
)
}
\arguments{
\item{modelType}{string, "A", "AD", "DirDom".
modelType="A": additive-only, GEBVS
modelType="AD": the "classic" add-dom model, GEBVS+GEDDs = GETGVs
modelType="DirDom":
the "genotypic" add-dom model with prop. homozygous
fit as a fixed-effect, to estimate a genome-wide inbreeding effect.
obtains add-dom effects, computes
allele sub effects (\eqn{\alpha = a + d(q-p)})
incorporates into GEBV and GETGV}

\item{selInd}{logical, TRUE/FALSE, selection index accuracy estimates,
requires input weights via \code{SIwts}}

\item{SIwts}{required if \code{selInd=FALSE}, named vector of selection
index weights, names match the "Trait" variable in \code{blups}}

\item{getMarkEffs}{T/F return marker effects, backsolved from GBLUP}

\item{returnPEV}{T/F return PEVs in GBLUP}

\item{blups}{nested data.frame with list-column "TrainingData" containing
BLUPs. Each element of "TrainingData" list, is data.frame with de-regressed
BLUPs, BLUPs and weights (WT) for training and test.}

\item{grms}{list of genomic relation matrices (GRMs, aka kinship matrices).
Any genotypes in the GRMs get predicted with, or without phenotypes.
Each element is named either A or D. Matrices supplied must match
required by A, AD and DirDom models. e.g. grms=list(A=A,D=D).}

\item{dosages}{dosage matrix. required only for modelType=="DirDom".
Also required if getMarkEffs==TRUE.
Assumes SNPs coded 0, 1, 2. Nind rows x Nsnp
cols, numeric matrix, with rownames and colnames to indicate SNP/ind ID}

\item{gid}{string variable name used for genotype ID's in e.g. \code{blups} (default="GID")}

\item{ncores}{number of cores}

\item{nBLASthreads}{number of cores for each worker to use for multi-thread BLAS}
}
\value{
tibble, one row, two list columns (basically a named two-element
list of lists): \code{gblups[[1]]} and \code{genomicPredOut[[1]]}.
code{gblups[[1]]}: tibble of predicted GEBV/GETGV, all traits and potentially
SELIND genomic BLUPs along the columns.

\code{genomicPredOut[[1]]} is a tibble that contains
some combination of lists-columns:
\itemize{
 \item gblups
 \item varcomps,
 \item fixeffs,
 \item allelesubsnpeff,
 \item addsnpeff,
 \item domstar_snpeff,
 \item domsnpeff
}
}
\description{
Run GBLUP model using \code{\link[sommer]{mmer}}, potentially on multiple
traits. Returns genomic BLUPs (GEBV and GETGV). If requested, returns
backsolved marker effects (equivalent to ridge regression / SNP-BLUP).
Three models are
enabled: additive-only ("A"), additive-plus-dominance ("AD") and a
directional-dominance model that incorporates a genome-wide homozygosity
effect ("DirDom"). Inbreeding effect is included in output GEBV/GETGV
predictions *after* backsolving SNP effects. If requested, returns
GEBV/GETGV computed for a selection index using \code{selInd=TRUE}
and supplying \code{SIwts}.
}
\seealso{
Other prediction_functions: 
\code{\link{predictCrosses}()}
}
\concept{prediction_functions}
