% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentWiseCrossVal.R
\name{runParentWiseCrossVal}
\alias{runParentWiseCrossVal}
\title{Run parent-wise cross-validation}
\usage{
runParentWiseCrossVal(
  nrepeats,
  nfolds,
  seed = NULL,
  modelType,
  ncores = 1,
  nBLASthreads = NULL,
  outName = NULL,
  ped = ped,
  gid = "GID",
  blups,
  dosages,
  grms,
  haploMat,
  recombFreqMat,
  selInd,
  SIwts = NULL,
  ...
)
}
\arguments{
\item{nrepeats}{number of repeats}

\item{nfolds}{number of folds}

\item{seed}{integer, make the parent trait-test folds reproducible}

\item{modelType}{string, "A", "AD", "DirDom". modelType="A": additive-only,
predicts mean and variances for breeding values [BVs]). modelType="AD": the
"classic" add-dom model, predicts family mean and variance for breeding
value based on allele sub. effects. Predicts family variance-covariances
for TGVs (BVs+DDs). Doesn't predict family mean TGV. modelType="DirDom":
the "genotypic" add-dom model with prop. homozygous fit as a fixed-effect,
to estimate a genome-wide inbreeding effect. obtains add-dom effects,
computes allele sub effects (\eqn{\alpha = a + d(q-p)}) predicts family
mean and covars for BVs using allele sub. effects predicts covars AND the
means for TGVs using directly the add-dom effects. estimated linear effect
of overall homozygosity (b), interpreted as inbreeding depression or
heterosis depending on its direction relative to each trait (Xiang et al.
2016). The estimated genome-wide fixed-effect of homozygosity (b) can be
incorporated into the predicted means and variances by first dividing by
the number of effects (p) and subtracting that value from the vector of
dominance effects (\eqn{d\ast}), to get \eqn{d=d*-\frac{b}{p}}}

\item{ncores}{number of cores}

\item{nBLASthreads}{number of cores for each worker to use for multi-thread BLAS}

\item{outName}{default=NULL (optional), name and path to save outputs}

\item{ped}{data.frame, 3 columns, "GID" (or \code{gid}), "sireID", "damID"
for male and female parent, respectively.}

\item{gid}{string variable name used for genotype ID's in e.g. \code{blups} (default="GID")}

\item{blups}{nested data.frame with list-column "TrainingData" containing}

\item{dosages}{dosage matrix. required only for modelType=="DirDom".
Assumes SNPs coded 0, 1, 2. Nind rows x Nsnp
cols, numeric matrix, with rownames and colnames to indicate SNP/ind ID}

\item{grms}{list of genomic relation matrices (GRMs, aka kinship matrices).
Any genotypes in the GRMs get predicted with, or without phenotypes.
Each element is named either A or D. Matrices supplied must match
required by A, AD and DirDom models. e.g. grms=list(A=A,D=D).}

\item{haploMat}{matrix of phased haplotypes, 2 rows per sample, cols = loci, {0,1}, rownames assumed to contain GIDs with a suffix, separated by "_" to distinguish haplotypes}

\item{recombFreqMat}{a square symmetric matrix with values = (1-2*c1), where c1=matrix of expected recomb. frequencies. The choice to do 1-2c1 outside the function was made for computation efficiency; every operation on a big matrix takes time.}

\item{selInd}{logical, TRUE/FALSE, selection index accuracy estimates,
requires input weights via \code{SIwts}}

\item{SIwts}{required if \code{selInd=FALSE}, named vector of selection
index weights, names match the "Trait" variable in \code{blups}}

\item{...}{}
}
\value{
tibble, one row, two list columns (basically a named two-element
list of lists): \code{meanPredAccuracy} and
\code{varPredAccuracy} both contain tibbles. Column "AccuracyEst" for
family-size weighted prediction accuracy estimates. If \code{selInd=TRUE}
then corresponding accuracy labelled "SELIND" in "Trait" columns.
The column "predVSobs" is a list of tibbles each containing the paired
predicted and observed values for the given repeat-fold-trait.
}
\description{
Assess the accuracy of predicted previously unobserved crosses.
Specifically the accurracy predicting the mean and variance among family
members in breeding and total genetic values. The cross-validation procedure
implemented is described in detail in the manuscript:
\url{https://www.biorxiv.org/content/10.1101/2021.01.05.425443v1}, see "Details" below.
User supplies a pedigree, haplotypes and other inputs.
}
\details{
First, define a vector, \eqn{\boldsymbol{P}} of the parents listed in the
pedigree. Define also a second vector \eqn{\boldsymbol{C}} listing the
genotypes (clones) in the pedigree, including the parents
(\eqn{\boldsymbol{P}\subset\boldsymbol{C}}).

Conducts \code{nrepeats} replications of the following procedure:

\enumerate{
     \item Define parent-wise cross-validation folds: randomly assign the
     parents in \eqn{\boldsymbol{P}} into \eqn{\textit{k}}-folds.
     \eqn{\boldsymbol{P}_{TST}^k}, the list of "test" parents in the
     \eqn{\textit{k}}th-fold.
     \item For each of the \emph{k}-folds (set of "test" parents), divide
     the clones vector \eqn{\boldsymbol{C}}into two mutually exclusive sets:
      "training" (\eqn{\boldsymbol{C}_{TRN}}) and "validation"
      (\eqn{\boldsymbol{C}_{VLD}}). From the set \eqn{\boldsymbol{C}_{TRN}},
      we exclude all descendants (offspring, grandchildren,
      great grandchildren, etc.) of \eqn{\boldsymbol{P}_{TST}^k}.
      We include the \eqn{\boldsymbol{P}_{TST}^k} themselves
      (phenotyping the parents before predicting their offspring) and any
      non-descendents. Define \eqn{\boldsymbol{C}_{VLD}} simply as the set
      difference between \eqn{\boldsymbol{C}} and \eqn{\boldsymbol{C}_{TRN}}.
     \item Estimate marker effects independently by fitting
     mixed-models (see section
below for further details) to \eqn{\boldsymbol{C}_{VLD}} and
\eqn{\boldsymbol{C}_{TRN}} corresponding to each
\eqn{\boldsymbol{P}_{TST}^k}.
     \item For each \eqn{\boldsymbol{P}_{TST}^k},
define the set of crosses to predict, \eqn{\boldsymbol{X}_{toPred}^k} to
include any of the 462 actual families (sire-dam pairs) in the pedigree, in
which the \eqn{\boldsymbol{P}_{TST}^k} were involved. By construction, the
real family members that have been observed for each of the
\eqn{\boldsymbol{X}_{toPred}^k} were excluded from the model used to get
marker effects for \eqn{\boldsymbol{C}_{TRN}}, and included in the model for
\eqn{\boldsymbol{C}_{VLD}}. Predict the means, variances and covariances for
each focal trait in each cross, \eqn{\boldsymbol{X}_{toPred}^k} using the
\eqn{\boldsymbol{C}_{TRN}} marker effects only.
     \item For each family in
\eqn{\boldsymbol{X}_{toPred}^k}, using all existing family members, compute
the sample means, variances and covariances for \strong{GEBV} and
\strong{GETGV} as predicted by the \eqn{\boldsymbol{C}_{VLD}} marker effects.
     \item Calculate the accuracy of prediction for each mean
(\eqn{\overset{\mu_{T}}{\textbf{cor}}_{BV}},
\eqn{\overset{\mu_{T}}{\textbf{cor}}_{TGV}}), variance
(\eqn{\overset{\sigma^2_{t=t}}{\textbf{cor}}_{BV}},
\eqn{\overset{\sigma^2_{t=t}}{\textbf{cor}}_{TGV}}) and covariance
(\eqn{\overset{\sigma_{t \neq t}}{\textbf{cor}}_{BV}},
\eqn{\overset{\sigma_{t \neq t}}{\textbf{cor}}_{TGV}})  in terms of both
\strong{BV }and \strong{TGV}. For \eqn{\overset{\mu_{T}}{\textbf{cor}}} we
used the Pearson correlation between predicted and sample mean
\strong{GEBV/GETGV}. For \eqn{\overset{\sigma^2_{t=t}}{\textbf{cor}}} and
\eqn{\overset{\sigma^2_{t\neq t}}{\textbf{cor}}}, only families with greater
than two members were able to be included, and we weighted the correlation
between the predicted and sample (co)variance of \strong{GEBV/GETGV}
according to the family size (R \strong{package}::\emph{function}
\strong{psych}::\emph{cor.wt}). For sake of comparison, we also include
accuracies in the supplement where predicted values are correlated to
phenotypic (rather than genomic-predicted) BLUPs, e.g.
\eqn{\overset{\mu_{T}}{\textbf{cor}}_{BV,BLUP}},
\eqn{\overset{\mu_{T}}{\textbf{cor}}_{TGV,BLUP}}, etc.
}
}
\seealso{
Other CrossVal: 
\code{\link{runCrossVal}()}
}
\concept{CrossVal}
