% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentWiseCrossVal.R
\name{varPredAccuracy}
\alias{varPredAccuracy}
\title{Title}
\usage{
varPredAccuracy(
  crossValOut,
  snpeffs,
  ped,
  modelType,
  selInd = FALSE,
  SIwts = NULL,
  ncores
)
}
\arguments{
\item{ncores}{}
}
\description{
Title
}
